LayoutTests/plugins/bindings-test.html

<html>
<head>
<script>
function doTest()
{
    if (window.layoutTestController)
        layoutTestController.dumpAsText();

    if (window.objCPlugin) {
        // Type should be "object"
        alert("typeof(objCPlugin)=" + typeof(objCPlugin));

        // Calling objCPlugin() should raise an exception.
        var result;
        try {
            result = objCPlugin();
        } catch (e) {
            result = e.name;
        }
        alert("objCPlugin()=" + result);
    }
}
</script>
</head>
<body onload="doTest();">
<p>This page tests invoking as functions JavaScript objects that are provided by plugins.</p>
<embed id="testCPlugin" type="application/x-webkit-test-netscape"></embed>
</body>
</html>
