LayoutTests/plugins/bindings-test-expected.txt

ALERT: typeof(objCPlugin)=object
ALERT: objCPlugin()=TypeError
This page tests invoking as functions JavaScript objects that are provided by plugins.
